Moved the last structures and enums containings `#[cfg(..)]` attributes to the auxiliary crate.
